warning hunt

+++ PostgreSQL Inspector/PostgreSQL Inspector/PGConstraint.m
@@ -34,7 +34,7 @@
      " order by c.oid ";
     
     PGCommand *command = [[PGCommand alloc] initWithConnection:connection commandText:commandText];
-    [command addParameter:[[PGOid alloc] initWithValue:relationId]];
+    [command addParameter:[[PGOid alloc] initWithValue:(unsigned int)relationId]];
     
     PGResult *result = [[command execute] objectAtIndex:0];
     NSMutableArray *constraints = [[NSMutableArray alloc] initWithCapacity:result.rows.count];
